+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Documentation for kraken now mentions the use-case: metering.
+- Documentation for metering now mentions the use-case: metering.
 - Fixed meshcloud-sso module output reference
 - Fixed syntax error in a resource in meshcloud-sso module
 - Add permissions for tags replication